requirements.txt
python-telegram-bot==21.0
openai==1.12.0
python-dotenv==1.0.0
aiohttp==3.9.0
ruff==0.1.14 